internmatch_backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0                  # async MongoDB driver
pymongo==4.7.2
beanie==1.26.0                # ODM for MongoDB / Motor
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0   # JWT
passlib[bcrypt]==1.7.4
python-multipart==0.0.9       # file uploads
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
aiofiles==23.2.1
python-dotenv==1.0.1
httpx==0.27.0                 # test client
pytest==8.2.0
pytest-asyncio==0.23.7
